Move GemQWebView into a separate file in the ui directory
